Remove rake, appraisal from gemspec executables
The contents of the bin directory are binstubs to assist development.
They are not intended to be shipped with the gem itself. This commit
updates the gemspec to ensure that they are not exposed to rubygems as
executables for redis-rack, which will fix conflicts with the legitimate
rake and appraisal executables provided by those other gems.