Raise Error when `:redis_store` is Incorrect Type
Passing an object that isn't a `Redis::Store` into the session storage
configuration can cause errors when methods are called with more
arguments than the default Redis client can handle. These additional
arguments are used by `Rack::Session::Redis` to configure the key
namespace and TTL globally in the main configuration, and pass them down
to `Redis::Store` to be appended to the key. An `ArgumentError` will now
be thrown when a `:redis_store` is of a type that isn't a `Redis::Store`
(or a subclass thereof).

Resolves #46