Set the TTL correctly on session keys
Previously, we were passing the whole of the `options` object to the
redis set command, which while not malicious, doesn't actually do
anything since none of those keys are supported. The only valid option
that is exposed in those options would be the expire_after key which we
need to transform into the `ex` option. If there's no expire_after then
the `ex` gets `nil` which then has the same behaviour as before.