Mitigate race condition (#51)
The session store is susceptible to a race condition when two different
processes attempt to update the same session. This is a realistic
situation because several popular Ruby web servers achieve parallelism
by running multiple instances in separate processes.

This commit mitigates (but does not entirely fix) this, by atomically
reading the existing content of the session before merging it the with
the updated session. This ensures that any data, which has been added to
the session in another concurrent request, does not get overwritten by
the current one.

It does not however address the situation where data has been deleted
from the session by a concurrent request. In that case we still risk
reintroducing data that was supposed to be gone. It's important to note,
however, that this is an existing issue, and not introduced by this
commit. Fixing is possible but requires some extra, non-trivial
bookkeeping (see comment on line 69 in lib/rack/session/redis.rb).